+++ CHANGELOG.md
@@ -55,11 +55,14 @@ and this project adheres to
 - @cosmjs/tendermint-rpc: Fix `key` and `value` type in `RpcAbciQueryResponse`
   to also include the `null` option.
 - @cosmjs/tendermint-rpc: Fix decoding events without attributes ([#1198]).
+- @cosmjs/amino, @cosmjs/proto-signing: Support amounts larger than the uint64
+  range in `parseCoins` ([#1268]).
 
 [#1170]: https://github.com/cosmos/cosmjs/issues/1170
 [#1177]: https://github.com/cosmos/cosmjs/issues/1177
 [#1188]: https://github.com/cosmos/cosmjs/pull/1188
 [#1198]: https://github.com/cosmos/cosmjs/pull/1198
+[#1268]: https://github.com/cosmos/cosmjs/issues/1268
 
 ### Changed